pyproject.toml
[tool.poetry]
name = "pyifcount"
version = "0.1.0"
description = ""
authors = ["FUKUDA Shoma <fkshom@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
pytest-mock = "^3.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
assertpy = "^1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
#addopts = "-v --capture=no -s"
#addopts = -s
log_cli = "1"
log_cli_level = "DEBUG"
log_cli_format = "%(message)s"
